requirements.txt
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2023.11.17
exceptiongroup==1.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
python-dotenv==1.0.0
python-telegram-bot==20.7
sniffio==1.3.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
